crawler/requirements.txt
lxml==4.8.0
pandas==1.3.5
pytorch==1.10.0
tqdm==4.64.0
transformers==4.16.2
tensorboardX==2.2
sklearn